+++ common/src/main/java/traben/entity_texture_features/mixin/MixinIdentifier.java
@@ -22,9 +22,9 @@ public abstract class MixinIdentifier {
             if (ETFConfigData.illegalPathSupportMode != ETFConfig.IllegalPathMode.None) {
                 if (!cir.getReturnValue() && path != null) {
 
-                    //noinspection EnhancedSwitchMigration
+
                     switch (ETFConfigData.illegalPathSupportMode) {
-                        case Entity: {
+                        case Entity -> {
                             if ((path.contains("/entity/") || path.contains("/optifine/") || path.contains("/etf/"))
                                     && (path.endsWith(".png") || path.endsWith(".properties") || path.endsWith(".mcmeta"))) {
                                 ETFUtils2.logWarn(ETFVersionDifferenceHandler.getTextFromTranslation("config.entity_texture_features.illegal_path_warn").getString()
@@ -37,16 +37,13 @@ public abstract class MixinIdentifier {
                                 cir.setReturnValue(true);
                             }
                         }
-                        break;
-                        case All:
+                        case All -> {
                             ETFUtils2.logWarn(ETFVersionDifferenceHandler.getTextFromTranslation("config.entity_texture_features.illegal_path_warn").getString()
                                     + " [" + path + "]");
                             if (!path.isBlank())
                                 cir.setReturnValue(true);
-                            break;
-                        default:
-                            ETFUtils2.logWarn("this message should not appear #65164");
-                            break;
+                        }
+                        default -> ETFUtils2.logWarn("this message should not appear #65164");
                     }
                 }
             }